Add usprawnienie działania funkcji add, edit, return

+++ Website/views.py
@@ -1,7 +1,6 @@
 from flask import Blueprint, render_template, request, flash, redirect, url_for
 from .models import Book, Author, User
 from .schemas import BookSchema
-import requests
 from . import db
 
 views=Blueprint('views', __name__)
@@ -21,18 +20,28 @@ def add():
     if request.method=='POST':
         title=request.form.get('title')
         authors=request.form.get('authors')
+        print(authors)
         available=request.form.get('available')
         for item in request.form.items():
             if not item[1]:
-                flash(f'missing reqired field "{item[0].capitalize()}"', category='error')
+                flash(f'missing required field "{item[0].capitalize()}"', category='error')
         if all(request.form.values()):
             new_book=Book(title=title, available=available)
-            author=Author(name=authors, book=title)
-            db.session.add(author)
+            books=Book.query.all()
+            if books:
+                book_id=books[-1].id+1
+            else:
+                book_id=1
+            for item in authors.strip().split(','):
+                author=Author(name=item.strip(), book=book_id)
+                if author not in [author.name for author in Author.query.all()]:
+                    db.session.add(author)
             db.session.add(new_book)
             db.session.commit()
             flash(f'book added (Title:{new_book.title})')
             books=Book.query.all()
+            for book in books:
+                print(book_id)
             books_list=[]
             for book in books:
                 books_list.append(BookSchema(book))
@@ -43,57 +52,70 @@ def add():
 
 def edit():
     if request.method=='POST':
+        book_id=request.form.get('id')
         new_title=request.form.get('title')
+        new_authors=request.form.get('authors')
         new_available=request.form.get('available')
-        book=Book.query.get(new_title)
+        book=Book.query.get(book_id)
         if book:
-            book=Book.query.get(new_title)
-            book.available=new_available
+            book=Book.query.get(book_id)
+            db.session.delete(book)
             db.session.commit()
-            flash(f'book updated ({new_title})')
+        new_book=Book(title=new_title, available=new_available)
+        db.session.add(new_book)
+        db.session.commit()
+        books=Book.query.all()
+        new_book=books[-1]
+        book_id=new_book.id
+        for item in new_authors.strip().split(','):
+            author=Author(name=item.strip(), book=book_id)
+            db.session.add(author)               
+        db.session.commit()
+        flash(f'book updated ({new_title})')
         return home()
     else:
         book=None
         if request.args:
-            book=BookSchema(Book.query.get(request.args['title']))
+            book=BookSchema(Book.query.get(request.args['id']))
         return render_template('edit.html', book=book)
     
 @views.route('/lend', methods=['GET', 'POST'])
 
 def lend():
     if request.method=='POST':
-        new_title=request.form.get('title')
+        new_id=request.form.get('id')
         new_available=request.form.get('available')
         new_user=request.form.get('user')
-        book=Book.query.get(new_title)
+        book=Book.query.get(new_id)
         if book:
-            book=Book.query.get(new_title)
-            user=User(name=new_user, book=new_title)
+            book=Book.query.get(new_id)
+            user=User(name=new_user, book=new_id)
             if new_user not in [user.name for user in User.query.all()]:
                 book.available=new_available
                 db.session.add(user)
                 db.session.commit()
-                flash(f'book lent ({new_title} -> {new_user})')
+                flash(f'book lent ({book.title} -> {new_user})')
             else:
                 flash('user already has another book')
         return home()
     else:
         book=None
         if request.args:
-            book=BookSchema(Book.query.get(request.args['title']))
+            book=BookSchema(Book.query.get(request.args['id']))
         return render_template('lend.html', book=book)
     
 @views.route('/return', methods=['GET', 'POST'])
 
 def return_book():
-    new_title=request.args['title']
+    new_id=request.args['id']
     new_user=request.args['user']
-    book=Book.query.get(new_title)
+    book=Book.query.get(new_id)
     if book:
-        book=Book.query.get(new_title)
+        book=Book.query.get(new_id)
         book.available='on'
-        user=User.query.get(new_user)
-        db.session.delete(user)
+        if new_user:
+            user=User.query.get(new_user)
+            db.session.delete(user)
         db.session.commit()
         flash('book returned')
     else:
